# Conflicts:
#	confignode/src/main/java/org/apache/iotdb/confignode/manager/consensus/ConsensusManager.java
#	node-commons/src/main/java/org/apache/iotdb/commons/service/ServiceType.java
#	server/src/main/java/org/apache/iotdb/db/conf/IoTDBDescriptor.java
#	server/src/main/java/org/apache/iotdb/db/mpp/plan/parser/StatementGenerator.java

+++ confignode/src/main/java/org/apache/iotdb/confignode/manager/load/balancer/region/CopySetRegionGroupAllocator.java
@@ -51,7 +51,7 @@ public class CopySetRegionGroupAllocator implements IRegionGroupAllocator {
   @Override
   public TRegionReplicaSet generateOptimalRegionReplicasDistribution(
       Map<Integer, TDataNodeConfiguration> availableDataNodeMap,
-      Map<Integer, Long> freeDiskSpaceMap,
+      Map<Integer, Double> freeDiskSpaceMap,
       List<TRegionReplicaSet> allocatedRegionGroups,
       int replicationFactor,
       TConsensusGroupId consensusGroupId) {
